chore: move useMemoizedFn to utils
